backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
psycopg2-binary==2.9.9
scikit-learn==1.4.0
numpy==1.26.3
requests==2.31.0
websockets==12.0
python-multipart==0.0.6
